feat: package-based server assignment, fleet deploy fixes, and MCC/MNC fleet sync
- Server Manager: servers categorized by package (development/starter/
  professional/enterprise), multiple servers per location (auto-suffixed
  ids), package field in deploy form + settings editor, package badge on
  server cards
- Registration: starter clients auto-assigned to the nearest least-loaded
  starter server via phone dialing-code region routing; professional/
  enterprise assigned manually by super admin; tenant editor filters
  assignable servers by the tenant's package
- Deploy: NOPASSWD sudo detection in deploy-to-server.sh (fixes installs
  on OVH boxes with passwordless sudo), new sydney-2 server registered
  and added to fleet deploys + IP lookup
- MCC/MNC: syncMccMncToTenants now only writes to locally-hosted tenant
  schemas (fleet replication fix), resync script for full re-propagation,
  canonical 2740-row dataset loaded on all 4 fleet servers
- Auth: per-account consecutive-failure login guards (tenant, super admin,
  webmail) with lockout
- DLR: pending delivery re-hydration on restart, DB-backed DLR queue
  persistence, timeout sweeper numeric casts
- Android SMS gateway, SMPP server/client hardening, K8s migration docs

+++ src/lib/dlr-timeout-sweeper.ts
@@ -316,7 +316,7 @@ export async function sweepTenantStaleDlrs(
 
       const upd = await client.query(
         `UPDATE messages SET dlr_status = 'DELIVERED', status = 'DELIVERED', dlr_timestamp = NOW(),
-         dlr_source = 'FORCE_TIMEOUT', cost = $2, supplier_cost = $3, profit = $2 - $3
+         dlr_source = 'FORCE_TIMEOUT', cost = $2::numeric, supplier_cost = $3::numeric, profit = $2::numeric - $3::numeric
          WHERE message_id = $1 AND dlr_status IN ('SENT', 'PENDING')`,
         [msg.message_id, updateCost, updateSuppCost]
       );
